Correct function types

+++ srcs/init.c
@@ -6,7 +6,7 @@
 /*   By: cempassi <cempassi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/25 15:09:07 by cempassi          #+#    #+#             */
-/*   Updated: 2021/06/25 16:33:18 by cempassi         ###   ########.fr       */
+/*   Updated: 2021/06/25 16:38:26 by cempassi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void set_defaults(t_nmap *nmap)
     nmap->ports = NULL;
 }
 
-int init_nmap(t_nmap *nmap, int argc, char **argv)
+uint8_t init_nmap(t_nmap *nmap, int argc, char **argv)
 {
     set_defaults(nmap);
     if (set_opts_args(nmap, argc, argv) != SUCCESS)
